Orders, redirects, fixes, etc.

+++ src/main/java/travelagency/dao/GenericDao.java
@@ -4,9 +4,9 @@ import java.util.List;
 
 public interface GenericDao<T> extends AutoCloseable {
     boolean create (T entity);
-    T findById(int id);
+    T findById(Long id);
     List<T> findAll();
     boolean update(T entity);
-    boolean delete(int id);
+    boolean delete(Long id);
     void close();
 }
